// src/api/generated/models/crimeTypeResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */

export interface CrimeTypeResponse {
  code?: string;
  name?: string;
  gaulaJurisdiction?: boolean;
}