// src/api/generated/models/kpiComparisonResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */
import type { KpiSummaryResponse } from './kpiSummaryResponse';
import type { ModalityComparisonResponse } from './modalityComparisonResponse';

export interface KpiComparisonResponse {
  currentTotal?: KpiSummaryResponse;
  previousTotal?: KpiSummaryResponse;
  byModality?: ModalityComparisonResponse[];
}